Use mt19937 to fix odd rand() bug
This may be just on my machine, but the circles are all in the top-left corner colliding constantly; you can't even see the circles because in every frame they collide. It seems it's rand's fault.
If I mod rng by RAND_MAX the bug comes back so maybe RAND_MAX is just too small.